update project clear

+++ app/Http/Controllers/PengajuanController.php
@@ -4,8 +4,10 @@ namespace App\Http\Controllers;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\PengajuanFormRequest;
+use App\Models\DetailAlat;
 use App\Models\pengajuan;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Date;
 
 class PengajuanController extends Controller
 {
@@ -14,27 +16,45 @@ class PengajuanController extends Controller
      */
     public function index()
     {
-
-        $data = pengajuan::orderby("id", "desc")->paginate(10);
+        $data = pengajuan::orderby("id", "desc")->with('detail_alat')->paginate(10);
         return view("admin-pengajuan.index")->with('data',$data);
     }
 
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request)
     {
-        return view("pengajuan.create");
+        $data = DetailAlat::where("pengajuan_id", $request->id)->with('peralatan')->get();
+        return view("admin-pengajuan.create", [
+            'data' => $data
+        ]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(PengajuanFormRequest $request)
+    public function store(Request $request)
     {
 
-        pengajuan::create($request->all());
-        return redirect()->route('pengajuan.index')->with('success', 'Berhasil Menambah Pengajuan');
+        $updatePengajuan = [
+            'total_penawaran' => $request->total_penawaran,
+            'status' => 202,
+            'tanggal_penawaran' => date("Y-m-d", strtotime(Date::now()))
+        ];
+
+        // dd($updatePengajuan);
+
+        for($i=0; $i< count($request->pengajuan_id); $i++) {
+            $updateDetailAlat = [
+                'harga' => $request->harga[$i],
+                'total_unit' => $request->total_unit[$i],
+                'updated_at' => date("Y-m-d H:i:s", strtotime(Date::now()))
+            ];
+            DetailAlat::where('id', $request->detail_id[$i])->update($updateDetailAlat);
+        }
+        pengajuan::where('id', $request->pengajuan_id[0])->update($updatePengajuan);
+        return redirect()->route('admin-pengajuan.index')->with('success', 'Berhasil melakukan penawaran');
     }
 
     /**
@@ -48,43 +68,44 @@ class PengajuanController extends Controller
     /**
      * Show the form for editing the specified resource.
      */
-    public function edit(string $id)
+    public function edit(Pengajuan $admin_pengajuan)
     {
-        $data = pengajuan::where('id', $id)->first();
-        return view('pengajuan.edit')->with('data', $data);
+        $data = DetailAlat::where("pengajuan_id", $admin_pengajuan->id)->with('peralatan')->get();
+        return view("admin-pengajuan.update", [
+            'data' => $data
+        ]);
     }
 
     /**
      * Update the specified resource in storage.
      */
-    public function update(PengajuanFormRequest $request, string $id)
+    public function update(Request $request, pengajuan $admin_pengajuan)
     {
-        // $request->validate([
-        //     'peralatan' => 'required',
-        //     'qty' => 'required',
-        //     'unit' => 'required',
-        //     'harga' => 'nullable'
+        $updatePengajuan = [
+            'total_penawaran' => $request->total_penawaran,
+            'status' => 202,
+            'tanggal_penawaran' => date("Y-m-d", strtotime(Date::now()))
+        ];
 
-            
-        // ], [
-        //     'peralatan.required' => 'Peralatan wajib diisi',
-        //     'qty.required' => 'QTY wajib diisi',
-        //     'unit.required' => 'Unit wajib diisi'
-        // ]);
-
-        $data = pengajuan::findOrFail($id);
-
-        $data->update($request->all());
-
-        return redirect()->route('pengajuan.index')->with('success', 'Berhasil Mengubah Pengajuan');
+        for($i=0; $i< count($request->pengajuan_id); $i++) {
+            $updateDetailAlat = [
+                'harga' => $request->harga[$i],
+                'total_unit' => $request->total_unit[$i],
+                'updated_at' => date("Y-m-d H:i:s", strtotime(Date::now()))
+            ];
+            DetailAlat::where('id', $request->detail_id[$i])->update($updateDetailAlat);
+        }
+        pengajuan::where('id', $request->pengajuan_id[0])->update($updatePengajuan);
+        return redirect()->route('admin-pengajuan.index')->with('success', 'Berhasil melakukan penawaran');
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $id)
+    public function destroy(Pengajuan $admin_pengajuan)
     {
-        pengajuan::where('id', $id)->delete();
-        return redirect()->to(route('pengajuan.index'))->with('delete', 'Berhasil Melakukan Delete Data');
+        pengajuan::where('id', $admin_pengajuan->id)->delete();
+        DetailAlat::where('id', $admin_pengajuan->id)->delete();
+        return redirect()->to(route('admin-pengajuan.index'))->with('delete', 'Berhasil Melakukan Delete Data');
     }
 }
